Refactor GenTcp.Client and GenTcp.Server
This commit refactors the `GenTcp.Client` and `GenTcp.Server` modules.

In `GenTcp.Client`, the `handle_call` function now uses the `hostname` parameter instead of `url` for connecting to the server. The `opts` variable is also introduced to set additional options for the connection. The `Logger.log` statements are updated to include more information about the socket.

In `GenTcp.Server`, the `network_loop` function now includes a `Logger.log` statement to indicate the start of the network loop. The `serve` function also includes a `Logger.log` statement to indicate that it is serving the socket.

These changes improve the readability and maintainability of the code.